# docs/swagger/swagger.yaml
basePath: /
...
definitions:
...
  github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.CreateFeeScheduleItemRequest:
    properties:
      name:
        example: interchange
        maxLength: 100
        type: string
      priority:
        example: 1
        type: integer
      structure:
        additionalProperties: {}
        type: object
      structureType:
        enum:
        - FLAT
        - PERCENTAGE
        - TIERED
        example: PERCENTAGE
        type: string
    required:
    - name
    - structure
    - structureType
    type: object
  github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.CreateFeeScheduleRequest:
    properties:
      applicationOrder:
        enum:
        - PARALLEL
        - CASCADING
        example: PARALLEL
        type: string
      currency:
        example: USD
        maxLength: 3
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.CreateFeeScheduleItemRequest'
        maxItems: 100
        minItems: 1
        type: array
      name:
        example: Card Processing - Visa
        maxLength: 100
        minLength: 1
        type: string
      roundingMode:
        enum:
        - HALF_UP
        - BANKERS
        - FLOOR
        - CEIL
        - TRUNCATE
        example: HALF_UP
        type: string
      roundingScale:
        example: 2
        type: integer
    required:
    - applicationOrder
    - currency
    - items
    - name
    type: object
...
  github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.FeeScheduleItemResponse:
    properties:
      createdAt:
        example: "2025-01-15T10:30:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: interchange
        type: string
      priority:
        example: 1
        type: integer
      structure:
        additionalProperties: {}
        type: object
      structureType:
        enum:
        - FLAT
        - PERCENTAGE
        - TIERED
        example: PERCENTAGE
        type: string
      updatedAt:
        example: "2025-01-15T10:30:00Z"
        type: string
    type: object
  github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.FeeScheduleResponse:
    properties:
      applicationOrder:
        enum:
        - PARALLEL
        - CASCADING
        example: PARALLEL
        type: string
      createdAt:
        example: "2025-01-15T10:30:00Z"
        type: string
      currency:
        example: USD
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.FeeScheduleItemResponse'
        maxItems: 100
        type: array
      name:
        example: Card Processing - Visa
        type: string
      roundingMode:
        enum:
        - HALF_UP
        - BANKERS
        - FLOOR
        - CEIL
        - TRUNCATE
        example: HALF_UP
        type: string
      roundingScale:
        example: 2
        type: integer
      tenantId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      updatedAt:
        example: "2025-01-15T10:30:00Z"
        type: string
    type: object
...
  github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.SimulateFeeItem:
    properties:
      baseUsed:
        example: "100.00"
        type: string
      fee:
        example: "1.50"
        type: string
      name:
        example: interchange
        type: string
    type: object
  github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.SimulateFeeRequest:
    properties:
      currency:
        example: USD
        type: string
      grossAmount:
        example: "100.00"
        type: string
    required:
    - currency
    - grossAmount
    type: object
  github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.SimulateFeeResponse:
    properties:
      currency:
        example: USD
        type: string
      grossAmount:
        example: "100.00"
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.SimulateFeeItem'
        maxItems: 100
        type: array
      netAmount:
        example: "97.70"
        type: string
      totalFee:
        example: "2.30"
        type: string
    type: object
...
  github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.UpdateFeeScheduleRequest:
    properties:
      applicationOrder:
        enum:
        - PARALLEL
        - CASCADING
        example: CASCADING
        type: string
      name:
        example: Updated Schedule
        maxLength: 100
        type: string
      roundingMode:
        enum:
        - HALF_UP
        - BANKERS
        - FLOOR
        - CEIL
        - TRUNCATE
        example: BANKERS
        type: string
      roundingScale:
        example: 4
        type: integer
    type: object
...
  http.ErrorResponse:
    description: Standard error response returned by all API endpoints
    properties:
      code:
        description: HTTP status code
        example: 400
        type: integer
      message:
        description: Human-readable error message
        example: context name is required
        type: string
      title:
        description: Error type identifier
        example: invalid_request
        type: string
    type: object
...
info:
  contact:
    email: support@lerian.studio
    name: Lerian Studio Support
    url: https://discord.gg/DnhqKwkGv3
  description: |-
    Reconciliation engine for the Lerian Studio ecosystem.
    Provides automated transaction matching between Midaz ledger and external systems.
  license:
    name: Lerian Studio General License
    url: https://github.com/LerianStudio/matcher/blob/main/LICENSE.md
  termsOfService: http://swagger.io/terms/
  title: Matcher Reconciliation API
  version: v1.0.0
paths:
...
  /v1/config/fee-schedules:
    get:
      description: Returns fee schedules for the tenant. Results may be capped by
        the "limit" query parameter (default 20, max 200).
      operationId: listFeeSchedules
      parameters:
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - default: 20
        description: Maximum number of records to return
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of fee schedules
          schema:
            items:
              $ref: '#/definitions/github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.FeeScheduleResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List fee schedules
      tags:
      - Configuration Fee Schedules
    post:
      consumes:
      - application/json
      description: Creates a new fee schedule for transaction fee calculation.
      operationId: createFeeSchedule
      parameters:
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Idempotency key for safe retries
        in: header
        name: X-Idempotency-Key
        type: string
      - description: Fee schedule creation payload
        in: body
        name: feeSchedule
        required: true
        schema:
          $ref: '#/definitions/github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.CreateFeeScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created fee schedule
          schema:
            $ref: '#/definitions/github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.FeeScheduleResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: 'Conflict: duplicate resource or idempotency key in progress'
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a fee schedule
      tags:
      - Configuration Fee Schedules
  /v1/config/fee-schedules/{scheduleId}:
    delete:
      description: Removes a fee schedule by ID.
      operationId: deleteFeeSchedule
      parameters:
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Schedule ID
        format: uuid
        in: path
        name: scheduleId
        required: true
        type: string
      responses:
        "204":
          description: Fee schedule successfully deleted
        "400":
          description: Invalid schedule ID format
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Fee schedule not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a fee schedule
      tags:
      - Configuration Fee Schedules
    get:
      description: Returns a fee schedule by ID.
      operationId: getFeeSchedule
      parameters:
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Schedule ID
        format: uuid
        in: path
        name: scheduleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved fee schedule
          schema:
            $ref: '#/definitions/github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.FeeScheduleResponse'
        "400":
          description: Invalid schedule ID format
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Fee schedule not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a fee schedule
      tags:
      - Configuration Fee Schedules
    patch:
      consumes:
      - application/json
      description: Updates fields on a fee schedule by ID.
      operationId: updateFeeSchedule
      parameters:
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Idempotency key for safe retries
        in: header
        name: X-Idempotency-Key
        type: string
      - description: Schedule ID
        format: uuid
        in: path
        name: scheduleId
        required: true
        type: string
      - description: Fee schedule updates
        in: body
        name: feeSchedule
        required: true
        schema:
          $ref: '#/definitions/github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.UpdateFeeScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated fee schedule
          schema:
            $ref: '#/definitions/github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.FeeScheduleResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Fee schedule not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: 'Conflict: duplicate resource or idempotency key in progress'
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a fee schedule
      tags:
      - Configuration Fee Schedules
  /v1/config/fee-schedules/{scheduleId}/simulate:
    post:
      consumes:
      - application/json
      description: Calculates fees for a given gross amount using a specific fee schedule.
      operationId: simulateFeeSchedule
      parameters:
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Schedule ID
        format: uuid
        in: path
        name: scheduleId
        required: true
        type: string
      - description: Simulation parameters
        in: body
        name: simulate
        required: true
        schema:
          $ref: '#/definitions/github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.SimulateFeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Simulation result
          schema:
            $ref: '#/definitions/github_com_LerianStudio_matcher_internal_configuration_adapters_http_dto.SimulateFeeResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Fee schedule not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Simulate fee calculation
      tags:
      - Configuration Fee Schedules
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication (format: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"